refactor(shared-admin): lift myyoast-connection slice out of ai-generator into shared-admin store
Moves the myyoast-connection Redux slice from packages/js/src/ai-generator/store/
to packages/js/src/shared-admin/store/ — the established home for cross-feature
slices — so that bulk-editor no longer couples to an unrelated ai-generator bundle.
Both ai-generator and bulk-editor now import from shared-admin/store directly.
