+++ packages/js/src/shared-admin/store/index.js
@@ -1,6 +1,7 @@
 export * from "./admin-url";
 export * from "./ai-generator-has-consent";
 export * from "./link-params";
+export * from "./myyoast-connection";
 export * from "./notifications";
 export * from "./plugin-url";
 export * from "./wistia-embed-permission";
